feat: add query to add a new pizza to the database (pizzas table update only)
Adding a new pizza to the database means editing more than one table:
pizzas, pizzas_categories, pizzas_ingredients

Hence a transaction is used, to ensure a consistend update and allowing
a rollback in case of errors.

Currently, only the pizzas table is updated.

The query is used in populatedb.js, to initialize the pizzas table based
on the data from datadb.json.