feat: add query to add a new category to the database (categories table update only)
Adding a new category to the database means editing more than one table:
categories, pizzas_categories, ingredients_incompatible_categories,
ingredients_enforced_categories

Hence a transaction is used, to ensure a consistend update and allowing
a rollback in case of errors.

Currently, only the categories table is updated.

The query is used in populatedb.js, to initialize the categories table
based on the data from datadb.json.