feat: add query to add a new ingredient to the database (ingredients table update only)
Adding a new ingredient to the database means editing more than one
table:
ingredients, pizzas_ingredients, ingredients_incompatible_categories,
ingredients_enforced_categories

Hence a transaction is used, to ensure a consistend update and allowing
a rollback in case of errors.

Currently, only the ingredients table is updated.

The query is used in populatedb.js, to initialize the ingredients table
based on the data from datadb.json.