feat: add multiple state checkbox ejs template
The possible states are passed as an array to this template (the first
state must be the unchecked state).

A single state is used as "name" property of the input element.

If states = ["unchecked","state1","state2","state3"], clicking on the
checkbox cycles through such states. On form sumbission, if you
instantiate multiple components of this kind, on the body object you
get: {state1: [checked elements], state2: [checked elements], ...}