feat: re-render the form on server-side validation errors
Since the errors occurs during the server-side validation, the same form
is rendered again to the user.

Note that the Post-Redirect-Get pattern is used only on successful
server-side validation.

Also, the matched data and error list (opportunely formatted) are passed
to the ejs template, but they are not used yet.