feat: add edit and delete actions link to the items views
Actions link to edit or delete a specific item (eg, a pizza, indentified
by its id) have been added to the pizzas views. The delete functionality
is not functional yet, while the edit one only displays the form to edit
a pizza, but the actual action is not performed, yet.

Note that the action links have been added to the categories and
ingredients views, too, although none of them is functional, yet.