feat: add route to edit an existing ingredient
This route renders the form page to edit an existing ingredient.

The form page is essentially the same as "ingredientNew" ejs view. Some minor
adjustments will be made later on (e.g., changing the page title), and
the handling of the post request will be adapted, too.

Currently, the page only shows the form filled with the data of the
ingredient to be edited.

Some db queries have been added to grab the proper data.