feat: shrink responsive tables only when they overflow horizontally
Previously, tables with responsive class were shrunk based on some media
query tied to the window width.

However, this does not take into account the fact that some tables might
need to shrink at larger or shorter widths based on the width of the
tables.

Now a js script is introduced to compute whether each table with
responsive class overflows horizontally, and in that case a shrink class
is applied, which is used to apply the proper styles without media
queries.