refactor: rewrite the update db queries for pizzas, ingredients, and categories
This reduces the number of db queries: eg, instead of deleting all
pizzas rules, re-inserting the updated one from the form, and then
compute which elements were updated, now first the items to update are
computed also based on the results of the queries to lock the updated
elements: then, only the items that are no more present are removed, and
those new ones are inserted.

Note that now the associated js functions return object { id, updated },
which can be used to determine whether the id is valid but not present
in the database (id would be undefined), which was not possible with
the previous version.