feat: create doughs table in the db
Only one dough type will be created, which will be the base dough used
for all the pizzas.

The pizza cost is given by the sum of this dough cost plus the cost of
each ingredient used in there.

The pizza availability is given by the minimum availability among the
dough and the used ingredients.

This allows for more realistic management of the pizzas.

In this implementation, the base dough will be made editable by the
admin, but not deletable. Also, it will not be possible to create new
doughs, as this is out of the scope of the app.