fix: show correct item name on title of edit pizza/ingredient/category/dough form page when validation errors
Before the fix, if a validation error occurred, the form page was
reloaded with the data from the compiled form. The page title used the
"name" field taken from the corresponding input element, hence if this
was edited, so did the item name in the title. Now the original name is
shown instead, which is more consistent since the item is not modified
yet in case of errors.