fix: handle the case in which ingredient / dough / pizzas cannot be restocked or ordered
This might happen when the ingredient / dough maximum stock limit is
reached or there are no available pizzas with the selected dough.

A custom error is shown, and the form input and submit button are
disabled. A custom error is thrown by the server in case a request
reaches the server.